Add spark plots to the daily report page.
In the table, performance differences are highlighted as a single percentage
difference.  Due to the noise in many benchmarks and boards, often the true
performance characteristic cannot be summarized well into a single percentage
number, and a human has to look at all the samples/data points collected by the
benchmarking runs to make a further assessment of whether a performance change
should be categorized as "noise" or "real".

To make that assessment, you need to look at the chart of performance data
points over time. By adding mini-charts with spark lines to the daily report
page, that human analysis can be done without having to open many separate
pages - speeding up that analysis massively.

This is implemented by adding SVG images to the daily report page, which means
that for the email report, this part probably won't show up in most email
clients. The only solution I can think of to also make it work for email
clients is to create images on the server - leading to more overloading of
the server.